# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authorization Application REST-API schema",
        "title": "Golang Authorization API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticates a user and creates a new session with access and refresh tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique idempotency key (UUID)",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully authenticated",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_romreign_AuthService_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_romreign_AuthService_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_romreign_AuthService_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_romreign_AuthService_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http.LoginRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "minLength": 3,
                    "example": "username"
                },
                "password": {
                    "type": "string",
                    "example": "superPass"
                }
            }
        },
        "internal_transport_http.LoginResponse": {
            "type": "object",
            "required": [
                "access_expired_at",
                "access_token",
                "refresh_expired_at",
                "refresh_token"
            ],
            "properties": {
                "access_expired_at": {
                    "type": "string"
                },
                "access_token": {
                    "type": "string",
                    "example": "4db118d8-a4df-438a-8128-24a9886393e5"
                },
                "refresh_expired_at": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "4db118d8-a4df-438a-8128-24a9886393e5"
                }
            }
        },
...
    }
}